tools/partner-app-integration/customer-app/requirements.txt
# Azure authentication and management
azure-identity>=1.15.0
azure-mgmt-authorization>=4.0.0
msal>=1.24.0

# Microsoft Graph API
msgraph-sdk>=1.0.0

# Configuration
python-dotenv>=1.0.0

# CLI and utilities
click>=8.1.0
rich>=13.0.0

# HTTP requests
requests>=2.31.0
